refactor routes and controllers
