+++ .coveralls.yml
@@ -1,0 +1,1 @@
+repo_token: DCx2Kq8SnhSBWOzMS8EasZjmKj7KdJ2nx
