+++ server/.env
@@ -1,0 +1,2 @@
+NODE_ENV='production'
+PORT='8000'
