--- server/src/controllers/center/allCenters.js
@@ -1,10 +1,0 @@
-import dbc from '../../data/dbc.json'
-
-const data = dbc.center
-
-class AllCenters {
-  static all(req, res) {
-    res.status(200).send(data)
-  }
-}
-export default AllCenters
