--- server/src/controllers/center/index.js
@@ -1,11 +1,0 @@
-import AllCenters from './allCenters'
-import AddCenter from './addCenter'
-import FindCenter from './findCenter'
-import UpdateCenter from './updateCenter'
-
-export {
-  AllCenters,
-  FindCenter,
-  AddCenter,
-  UpdateCenter
-}
