--- server/src/controllers/events/index.js
@@ -1,9 +1,0 @@
-import CreateEvent from './createEvent'
-import UpdateEvent from './updateEvent'
-import DeleteEvent from './deleteEvent'
-
-export {
-  CreateEvent,
-  UpdateEvent,
-  DeleteEvent
-}
