--- template/README.md
@@ -1,20 +1,0 @@
-# event-management
-Given you manage an events center, this app will help you accept applications to use your center / facilities, and will either decline events when the proposed day is already taken, or suggest an available day
-
-
-https://spectrumsun.github.io/eventmanager/
-
-
-i. User​ ​signup​ ​and​ ​signin​ ​pages.  
-ii. A​ ​page​ ​where​ ​an​ ​authenticated​ ​user​ ​can​ ​add​ ​a​ ​new​ ​event.  iii. A​ ​page,​ ​section​ ​or​ ​view​ ​where​ ​an​ ​authenticated​ ​user​ ​can   	1. Modify​ ​the​ ​event​ ​he/she​ ​added  
-	2. Delete​ ​the​ ​event​ ​he/she​ ​added  
-iv. A​ ​page​ ​where​ ​an​ ​admin​ ​can​ ​add​ ​a​ ​new​ ​center  
-v. A​ ​page,​ ​section​ ​or​ ​view​ ​where​ ​an​ ​admin​ ​can​ ​modify​ ​the​ ​details​ ​of​ ​a​ ​center  
-vi. A​ ​page​ ​showing​ ​the​ ​details​ ​of​ ​a​ ​center​ ​and​ ​the​ ​events​ ​slated​ ​for​ ​that  center
-
-
-Tecnoloogy Used
-HTML 5
-CSS3
-JavaScript
-Boot Strap 4
